+++ modules/client/CHANGELOG.md
@@ -210,7 +210,7 @@ On 2022-11-22 16:45:52
 - Automatic generated README from examples
 - Split clients into modules
 - Removed `javascript` folders
-- Updates `@aragon/core-contracts-ethers` to `v0.4.1-alpha`
+- Updates `@aragon/osx-ethers` to `v0.4.1-alpha`
 - Parameters types for DAO creation
 - Make `tokenAdresses` optional on `getBalances`
 ### Added
@@ -371,12 +371,12 @@ On 2022-06-14 15:03:15
 ## [0.0.13-alpha] - Tue May 24 09:26:45 UTC 2022  
 
 ### Added
-- Getting `DAOFactory` address from `@aragon/core-contracts-ethers` package depending on the selected network in `Context`.
+- Getting `DAOFactory` address from `@aragon/osx-ethers` package depending on the selected network in `Context`.
   
 ## [0.0.12-alpha] - Mon May 23 11:17:16 UTC 2022  
 
 ### Changed
-- Upgraded `@aragon/core-contracts-ethers` package to the `0.2.1-alpha` version and adapted the base code to accept the new `VoteConfigStruct` struct.
+- Upgraded `@aragon/osx-ethers` package to the `0.2.1-alpha` version and adapted the base code to accept the new `VoteConfigStruct` struct.
 
 ## [0.0.11-alpha] - Wed May 18 14:51:02 UTC 2022